run prettier/linter on reflection-build

+++ packages/reflection-build/test/examples/source-repository/a/symlinkTypescriptParser.js
@@ -5,9 +5,8 @@ const modulesPath = path.resolve('../../../../modules');
 const targetPath = path.resolve('./modules');
 
 try {
-  if (fs.existsSync(targetPath))
-    fs.unlinkSync(targetPath);
-  
+  if (fs.existsSync(targetPath)) fs.unlinkSync(targetPath);
+
   fs.symlinkSync(modulesPath, targetPath, 'junction'); // Use 'junction' for directories on Windows
 } catch (error) {
   console.error('Failed to create symlink:', error);
